refactor(vertical-menu): move context into an internal directory
Moves `VerticalMenuFullScreenContext` and `MenuItemContext` into `__internal__` directories in
their respective components and updates imports where neccessary. The contexts are not intended
for public use and should improve tree-shaking in Carbon.

BREAKING CHANGE: moves `VerticalMenuFullScreenContext`and `MenuItemContext` into
internal directories
