+++ src/components/vertical-menu/vertical-menu-full-screen/__internal__/vertical-menu-full-screen.context.ts
