+++ src/components/vertical-menu/vertical-menu-item/__internal__/menu-item.context.ts
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default React.createContext({
+  level: 0,
+});
